+++ lib-filesystem-view-fx/src/main/java/de/bernd_michaely/common/filesystem/view/fx/FileSystemTreeViewFX.java
@@ -1,11 +1,24 @@
-/* Created on Sep 14, 2021 */
+/*
+ * Copyright 2025 Bernd Michaely (info@bernd-michaely.de).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package de.bernd_michaely.common.filesystem.view.fx;
 
 import de.bernd_michaely.common.filesystem.view.base.Configuration;
 import de.bernd_michaely.common.filesystem.view.base.PathView;
 import de.bernd_michaely.common.filesystem.view.base.RootNodeCtrl;
 import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.SortedSet;
 import javafx.application.Platform;
@@ -14,12 +27,9 @@ import javafx.beans.property.ReadOnlyBooleanWrapper;
 import javafx.beans.property.ReadOnlyObjectProperty;
 import javafx.beans.property.ReadOnlyObjectWrapper;
 import javafx.collections.ListChangeListener.Change;
-import javafx.scene.control.TreeCell;
 import javafx.scene.control.TreeItem;
 import javafx.scene.control.TreeView;
 import javafx.scene.layout.Region;
-import javafx.scene.text.Font;
-import javafx.scene.text.FontPosture;
 import org.checkerframework.checker.nullness.qual.*;
 
 /**
@@ -33,44 +43,6 @@ class FileSystemTreeViewFX implements FileSystemTreeView
 	private final ReadOnlyBooleanWrapper pathSelectedProperty;
 	private final TreeView<PathView> treeView;
 	private final RootNodeCtrl rootNodeCtrl;
-
-	private static class PathViewTreeCell extends TreeCell<PathView>
-	{
-		private @MonotonicNonNull Font fontDefault;
-		private @MonotonicNonNull Font fontItalic;
-
-		private PathViewTreeCell(TreeView<PathView> treeView)
-		{
-		}
-
-		@SuppressWarnings("argument")
-		private void clear()
-		{
-			setText(null);
-			setGraphic(null);
-		}
-
-		@Override
-		protected void updateItem(PathView item, boolean empty)
-		{
-			super.updateItem(item, empty);
-			if (empty || item == null)
-			{
-				clear();
-			}
-			else
-			{
-				if (fontDefault == null || fontItalic == null)
-				{
-					fontDefault = getFont();
-					fontItalic = Font.font(
-						fontDefault.getFamily(), FontPosture.ITALIC, fontDefault.getSize());
-				}
-				setFont(Files.isSymbolicLink(item.getPath()) ? fontItalic : fontDefault);
-				setText(item.toString());
-			}
-		}
-	}
 
 	FileSystemTreeViewFX(Configuration configuration)
 	{
@@ -102,9 +74,7 @@ class FileSystemTreeViewFX implements FileSystemTreeView
 					}
 				}
 			});
-		this.treeView.setCellFactory(PathViewTreeCell::new);
-		// TODO:
-		//this.treeView.setCellFactory(TreeCellPathView::createInstance);
+		this.treeView.setCellFactory(TreeCellPathView::factory);
 	}
 
 	@Override