+++ lib-filesystem-view-fx/src/main/java/de/bernd_michaely/common/filesystem/view/fx/NodeViewFX.java
@@ -1,4 +1,18 @@
-/* Created on Mar 29, 2022 */
+/*
+ * Copyright 2025 Bernd Michaely (info@bernd-michaely.de).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package de.bernd_michaely.common.filesystem.view.fx;
 
 import de.bernd_michaely.common.filesystem.view.base.NodeView;
@@ -55,7 +69,7 @@ class NodeViewFX implements NodeView
 	public void setLeafNode(boolean leafNode)
 	{
 		treeItem.setLeaf(leafNode);
-		logger.log(TRACE, "SET FOR NODE »" + treeItem.getValue() + "« LEAF TO »" + leafNode + "«");
+		logger.log(TRACE, () -> "SET FOR NODE »" + treeItem.getValue() + "« LEAF TO »" + leafNode + "«");
 	}
 
 	TreeItem<PathView> getTreeItem()
@@ -77,7 +91,7 @@ class NodeViewFX implements NodeView
 				}
 				else
 				{
-					logger.log(WARNING, getClass().getName() +
+					logger.log(WARNING, () -> getClass().getName() +
 						"::insertSubNodeAt : Invalid NodeView : " + subNodeView);
 				}
 			});