+++ lib-filesystem-view-swing/src/main/java/de/bernd_michaely/common/filesystem/view/swing/JFileSystemTreeViewImpl.java
@@ -1,4 +1,18 @@
-/* Created on Sep 20, 2021 */
+/*
+ * Copyright 2025 Bernd Michaely (info@bernd-michaely.de).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package de.bernd_michaely.common.filesystem.view.swing;
 
 import de.bernd_michaely.common.filesystem.view.base.Configuration;
@@ -7,7 +21,6 @@ import java.awt.Component;
 import java.awt.Font;
 import java.io.IOException;
 import java.lang.System.Logger;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
@@ -128,7 +141,8 @@ class JFileSystemTreeViewImpl implements JFileSystemTreeView
 			{
 				final var pathView = treeNode.getUserObject();
 				final var path = pathView.getPath();
-				final boolean isSymbolicLink = Files.isSymbolicLink(path);
+				final var symbolicLinkTarget = pathView.getSymbolicLinkTarget();
+				final boolean isSymbolicLink = symbolicLinkTarget != null;
 				if (this.fontDefault == null)
 				{
 					this.fontDefault = getFont();
@@ -137,24 +151,8 @@ class JFileSystemTreeViewImpl implements JFileSystemTreeView
 				{
 					this.fontSymLink = fontDefault.deriveFont(Font.ITALIC);
 				}
-				String toolTipText;
-				if (isSymbolicLink)
-				{
-					try
-					{
-						toolTipText = pathView + " → " + Files.readSymbolicLink(path);
-					}
-					catch (UnsupportedOperationException | IOException ex)
-					{
-						toolTipText = null;
-					}
-				}
-				else
-				{
-					toolTipText = null;
-				}
 				setFont(isSymbolicLink ? fontSymLink : fontDefault);
-				setToolTipText(toolTipText);
+				setToolTipText(isSymbolicLink ? pathView + " → " + symbolicLinkTarget : null);
 			}
 			return treeCellRendererComponent;
 		}